+++ src/Domain/Model/EventTranslation.php
@@ -66,6 +66,12 @@ class EventTranslation
     /** @var string|null */
     private $mobileLogoExtension;
 
+    /** @var string|null */
+    private $notificationImage;
+
+    /** @var string|null */
+    private $notificationImageExtension;
+
     /**
      * @param Event       $event
      * @param string      $locale
@@ -267,15 +273,35 @@ class EventTranslation
         return $this->mobileLogoExtension;
     }
 
+    /**
+     * @return null|string
+     */
+    public function getNotificationImage(): ?string
+    {
+        return $this->notificationImage;
+    }
+
+    /**
+     * @return null|string
+     */
+    public function getNotificationImageExtension(): ?string
+    {
+        return $this->notificationImageExtension;
+    }
+
     public function updateLogoAndMobileLogo(
         ?string $logo = null,
         ?string $logoExtension = null,
         ?string $mobileLogo = null,
-        ?string $mobileLogoExtension = null
+        ?string $mobileLogoExtension = null,
+        ?string $notificationImage = null,
+        ?string $notificationImageExtension = null
     ): void {
         $this->logo = $logo;
         $this->logoExtension = $logoExtension;
         $this->mobileLogo = $mobileLogo;
         $this->mobileLogoExtension = $mobileLogoExtension;
+        $this->notificationImage = $notificationImage;
+        $this->notificationImageExtension = $notificationImageExtension;
     }
 }