+++ src/Infrastructure/Adapter/EventUrlGenerator.php
@@ -45,6 +45,16 @@ class EventUrlGenerator implements EventUrlGeneratorInterface
         );
     }
 
+    public function generateImageAbsoluteUrl(Event $event, string $path)
+    {
+        return sprintf(
+            '%s://%s%s',
+            $this->scheme,
+            $event->getDomain(),
+            $path
+        );
+    }
+
     public function generateBaseEventAbsoluteUrl(Event $event): string
     {
         return sprintf(