+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - VIMEET-2096 - Add image header to webinar
+- VIMEET-2050 - Can define an image to the header notification
 
 ## [2.26.0] - 2020-05-05
 ### Added